updated Makefile to make bin directory

+++ include/constant-rate/globalstructs.h
@@ -23,7 +23,7 @@
 
 struct GlobalParameters
 {
-  int tot_life;
+  double tot_life;
   int max_pop;
   int ancestors;
   int ancestor_clones;
