+++ include/time-dependent/globalstructs.h
@@ -30,7 +30,7 @@ typedef double (*RateFunctionsPtr) (double, void*);
 
 struct GlobalParameters
 {
-  int tot_life;
+  double tot_life;
   int max_pop;
   int ancestors;
   int ancestor_clones;
